* added new SC.CheckboxView and SC.RadioView for creating simulated checkbox and radio buttons.  These buttons have more functions than their platform native counterparts including mixed style and full styling support.
* Moved field view subclasses to views/field directory.

* Added SC.Validatable mixin that can be applied to any view to make it work
with validators.  SC.FieldView now incorporates this mixin.

* Now if you add initMixin() to a mixin, it will be called automatically 
whenever an object that includes the mixin is instantiated.  You can use this
to perform any setup necessary for the mixin.

* Renamed SC.Control.contentValueProperty to SC.Control.contentValueKey to
be consistent with the rest of the API.

* Added updateContentWithValueObserver to SC.Control that will relay changes
to the value property back to the content object if you use the default 
implementation.

* Changed SC.Control observers for various states that were once private to 
public methods.  This way you can override them in your subclasses if you need
to perform some extra processing and you want to make sure you do them in the 
right order.  You can also just add more observers as well.

* Added JSDoc info for Validators, SC.FieldView, SC.CheckboxFieldView



git-svn-id: https://sproutcore.googlecode.com/svn/trunk/samples@32 394f415c-e224-0410-a11c-cb241aa5d150